REFACTOR: move classes to packagge with correct name
