+++ src/main/java/segelzwerg/sporttooolbox/calculators/SpeedCalculator.java
@@ -1,9 +1,9 @@
 package segelzwerg.sporttooolbox.calculators;
 
 import lombok.EqualsAndHashCode;
-import segelzwerg.sporttooolbox.IUnits.Distance;
 import segelzwerg.sporttooolbox.IUnits.Pace;
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Distance;
+import segelzwerg.sporttooolbox.iunits.Time;
 import segelzwerg.sporttooolbox.iunits.speed.Speed;
 
 /**
