+++ src/main/java/segelzwerg/sporttooolbox/calculators/TimeCalculator.java
@@ -1,8 +1,8 @@
 package segelzwerg.sporttooolbox.calculators;
 
 import lombok.EqualsAndHashCode;
-import segelzwerg.sporttooolbox.IUnits.Distance;
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Distance;
+import segelzwerg.sporttooolbox.iunits.Time;
 import segelzwerg.sporttooolbox.iunits.speed.Speed;
 
 @EqualsAndHashCode
