+++ src/main/java/segelzwerg/sporttooolbox/iunits/Time.java
@@ -1,6 +1,8 @@
-package segelzwerg.sporttooolbox.IUnits;
+package segelzwerg.sporttooolbox.iunits;
 
 import lombok.EqualsAndHashCode;
+import segelzwerg.sporttooolbox.IUnits.MinutesPerKilometer;
+import segelzwerg.sporttooolbox.IUnits.Pace;
 import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 import segelzwerg.sporttooolbox.iunits.speed.Speed;
