+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/MeterPerSecond.java
@@ -3,7 +3,7 @@ package segelzwerg.sporttooolbox.iunits.speed;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Time;
 
 /**
  * Speed in meter per second
