+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/Speed.java
@@ -1,6 +1,6 @@
 package segelzwerg.sporttooolbox.iunits.speed;
 
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Time;
 
 /**
  * Speed interface
