+++ src/test/java/segelzwerg/sporttooolbox/calculators/TimeCalculatorFactoryTest.java
@@ -1,7 +1,7 @@
 package segelzwerg.sporttooolbox.calculators;
 
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.IUnits.Distance;
+import segelzwerg.sporttooolbox.iunits.Distance;
 import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 import segelzwerg.sporttooolbox.iunits.speed.Speed;
 import segelzwerg.sporttooolbox.web.speed.SpeedForm;
