+++ src/test/java/segelzwerg/sporttooolbox/calculators/TimeCalculatorTest.java
@@ -1,8 +1,8 @@
 package segelzwerg.sporttooolbox.calculators;
 
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.IUnits.Distance;
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Distance;
+import segelzwerg.sporttooolbox.iunits.Time;
 import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 
 import static org.hamcrest.MatcherAssert.assertThat;
