+++ src/test/java/segelzwerg/sporttooolbox/web/speed/SpeedServiceTest.java
@@ -2,7 +2,7 @@ package segelzwerg.sporttooolbox.web.speed;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.Time;
 import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 import segelzwerg.sporttooolbox.iunits.speed.Speed;
